Create change
Patch-set: 1
Change-id: I2f98a193aae500f533ca0504b1de026a979d7d06
Subject: In function SendResponseRaw a mutex is acquired before writing to a file; if the write operation fails the function returns without releasing the mutex, which will lead to a deadlock.  The fix ensures that the mutex is always released, regardless of the result of the write operation.
Branch: refs/heads/master
Commit: 2f98a193aae500f533ca0504b1de026a979d7d06