RIL API to support OMH
RIL APIs to support the Open Market Handset initiative. This provides the
flexibility for the operators/networks to define profiles for making data
calls. These profiles are for different service types (e.g. MMS, LBS etc) and
their associated priorities.

The APIs query the modem for profiles for a particular service type and retrieves all the data profiles and their priorities.

Change-Id: I11d378a4fe401ba61d259d0fce723757a073ae9b