The Android.mk file in hardware/ril/mock-ril defines some variables if and only if the architecuture is arm based. It then has some rules outside the arm "if" statement which uses these variables.
This causes problems when building on a Mac for non-arm targets because the find command in Mac OS mush have a base directory specified (i.e. find dir -name xxx), but due to the missing variables a find is attempted without a base directory (i.e. find -name xxx) which results in an error and the build subsequently failing with an error.

Change-Id: Ie588136e6c39ebff0f5cee04b7f44a4373bc3ca9