mock-ril: fix compiling issues on Mac OS X
The $(src_java) and $(gen_src_java) are empty in mock-ril/Android.mk
if the architecuture is not arm, or the build is non-debug variants.
It results in a "find  -name ..." statement with an empty searching
directory. BSD derived find on Mac OS X complains about that.

The patch tried to fix the issue in the ways:

* always defines $(src_java) and $(gen_src_java) no matter what TARGET_ARCH is.
* only defines module librilproto-java in debug variants build.

Change-Id: I8a04657d280abc8df1fc3cf95d7143c2225f6ef5

Signed-off-by: Chih-Wei Huang <cwhuang@linux.org.tw>