Update patch set 8
Patch Set 8:

Wink - what is the intention of keeping _V5 and _V6 on the structures (like for SIM_IO or Card_Status)? Will you be removing the _V5 and _V6 suffix before final HC release or is the intention to keep the old structure and have _V6 for the new one? e.g. RIL_SIM_IO and RIL_SIM_IO_V6 - so the first one is for backward compatibility and _V6 is for HC onwards?

Patch-set: 8
Reviewer: Vidya Rao <1004214@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0