Update patch set 8
Patch Set 8:

Answers to Vidya's questions:

"What is the intention of keeping _V5 and _V6 on the structures" - A: I want to be sure structures that have new formats won't compile so as to minimize accidentally returning old structures. By changing the name of both the old and new version it's easier to identify commands that have changed.

"Will you be removing the _V5/V6 suffix before final HC" - A: No

"Why not keep the old name unchanged and just add a _V6 name to new structures" - A: As I said above, I want to highlight the differences, trying to have one header file provide support for all versions is not the goal. Further more, we will be deleting old versions as time goes along.

Patch-set: 8
Label: Verified=0