Update patch set 12
Patch Set 12:

Hi Wink.

I think that some of the overall design decision for LTE
should be discussed and decided before we can iron out
all the details for this API.

I have written a document on LTE and VoLTE, discussing the
different option on how to handle the Initial PDN in LTE:

https://docs.google.com/viewer?a=v&pid=explorer&chrome=true&srcid=0B6TxXYDwShpjNDdhMzI5YzctNDhhMC00NGQ0LWE4MzgtYTUzZjMzODk3YzNi&hl=en

I have also included an VoLTE/IMS API.

Feedback on this is very much welcome. Particularly I'd
like to discuss how the Initial PDN should be handled in
the Android RIL.

Regards,
Sjur

Patch-set: 12