Update patch set 2
Patch Set 2: (1 inline comment)

The power varies depending on the exact radio technology. If we set the actual transmit power from telephony, telephony will need to know the actual radio technology and change it when there is a handoff.

If we leave that intelligence to the modem, it will be handle the FCC regulations for all radio technologies and telephony will only need to tell the modem/ril that wifi hotspot is enabled.

Patch-set: 2