Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(4 inline comments)

As usual when adding new ril commands it must be done in such a way as to allow for backwards compatibility. See https://review.source.android.com/#/c/20467 for techniques we've used when we went to ril.h version 6.

Also, RIL.java needs to be extended to support the new commands being introduced and I don't see the code that uses these new commands, but probably just missed it.

Patch-set: 2