Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

Took a quick look.

Is there a design document that you can share on the outlines the problem and the solution and the ramifications to the Framework and Applications?

This patch seems to do two things add some subscription commands and support for "multiple RILD", can they be separated?

What changes in the frame work are needed to support multiple RILD's?

This change looks to support only two RILD's if we're going to support multiple I'd want to be able to support N and I want it to gracefully fail if more than N are attempted. Also, we need to report the number of RILD's that can be supported, I'd like this to be dynamic.

Also, I want to be able to enumerate the RILD's that are currently running.

Please fix long lines?

Please use "ALOGE" not "LOGE" I see you've used both.

Patch-set: 1
Label: Code-Review=-1
Label: Verified=0