Update patch set 4
Patch Set 4:

Hi Jake,

Without this change to add type information, the current implemntation is this:

If pdu len is <= 56 && message id is of ETWS

    Decode as ETWS primary.

Else if pdu len is <= 88

    Decode as GSM format

Else // umts format guranteeded  >= 90

    Decode as UMTS format

Since ETWS 2ndary format can be 1-9600 octet sper TS36.413 9.2.1.53, that will break above assumptions.  And without the type information, we cannot determine what format to use for decoding in frameworks layer, and this is a huge problem.  Please review the change.

Patch-set: 4
Label: Code-Review=0