Add support for retrieving the Data Radio Resource Control State.
To be able reduce power consumption caused by excessive use
of the data radio in a high power state we'd like the radio to
report its current state the Framework. This will be used to
initially collect data and in the future reduce power consumption
by understanding the relationship between power, state and
data transmissions.

Change-Id: I3538f9378f066439b82bda630dea7d1b039eb54a