Update patch set 5
Patch Set 5:

The framework will use the information to try to reduce the power consumption by sending as much data as possible when when the radio is already in a high power state.

We've considered having a "go dormant" command but operators don't like that, so we haven't pushed it. But we do recommend that radio's implement fast dormancy.

Patch-set: 5