Notify RIL that the device is shutting down using RIL_REQUEST_RADIO_POWER
When device is in airplane mode, SIM card may be ON. Send a
RIL_REQUEST_RADIO_POWER with -1 to RIL to notify RIL that the
device is shutting down. RIL can use this request to shutdown the
SIM card and modem gracefully if SIM is left ON in airplane mode.

Bug: 9773278
Change-Id: I71ddfc996fb8eb2b1d2fecb9231190ef8c92f079