Update patch set 1
Patch Set 1: Code-Review-2

Marking as -2 so we don't accidentally submit.

This is the direction I want to go, new interfaces for new features instead of adding more and more to RIL. But I'd like to use protobuf as they are extensible and marshalling/unmarshalling is autogenerated, thus eliminating the need for that to be done by the daemon.

Is this totally objectionable to anyone?

Patch-set: 1
Reviewer: Wink Saville <1001401@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-2