Update patch set 1
Patch Set 1:

After some reflection, I think in the short term we probably should define a system property (or config.xml variable)  with a list of "rils" that TelephonyDevController would retrieve to get the initial configuration.  And then have not only the RIL_UNSOL_HARDWARE_CONFIG_CHANGED but also a RIL_GET_HARDWARE_CONFIG so that when TelephonyDevController has the list of rils it can fetch the current hardware configuration.

Your thoughts?

Patch-set: 1