Update patch set 1
Patch Set 1:

yeah, so the next thing to tackle is indeed initial config...  was thinking about this and how it would work.  some overlay.xml seems appropriate, the on-request is a good idea, but it would only work after the RIL gets instantiated - so after registerRIL - so registerRIL triggers a on-request which supersedes whatever overlay.xml value was found initially?

Patch-set: 1
Label: Verified=0