Update patch set 2
Patch Set 2: Code-Review-1

(2 comments)

So as to easily track the addition, removal or change of modem/sim configuration I think it might be a good idea to return the current list as an array. Although this causes redundant data it means there is no chance that a configuration change is miss interpreted as all information is available at the same time.

This is what we've done for RIL_UNSOL_DATA_CALL_LIST_CHANGED and its worked reasonably well for keeping track of the data call list unambiguously. See DcController, the code is "tricky" it has worked out.

Of course, you may have a better idea, I'm all ears.

Patch-set: 2
Label: Code-Review=-1