Update patch set 2
Patch Set 2:

i thought about the array, but i did not think it would be needed so long we have proper serialization and locking.  which i have put on the TDC side.

but, if we want to 're-use' an existing model that worked, we can certainly do the 'call list' equivalent.

Patch-set: 2
Reviewer: Pierre Couillaud <1025044@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0