Cleanup DATA_REG and VOICE_REG Location Info
The data registration and voice registration states contain
large swaths of location information. While it seemed at first
blush we could remove that information, having synchronized
knowledge of when a registration has resulted in a cell change
provides valuable infomation. Thus, I suggest that we perform
some cleanup by using the same cell location structure that is
being used elsewhere in the interface to ensure that this method
always keeps sufficiently up-to-date information, without the
need to increasingly bloat it. Note this also fixes subtle bugs
that existed due to the information previously being incomplete.

Bug: 32223344
Test: none
Change-Id: I45ae98b6708bd30d21d101025e083ecf46b7709a