Update patch set 2
Patch Set 2: Code-Review+1

Still lgtm. It was just a question, only concerned that if you touch or move it, you own it and then making it correct(?) makes sense is all. Is there an issue being addressed here (if this was internal, I would be looking for a bug number to discern context)? The commit message indicates what you are cleaning up, but update from what.

Naveen has more experience in this tree than I, consider me a neophyte.

Patch-set: 2
Label: Code-Review=+1