Update patch set 4
Patch Set 4:

Please rebase (!) gerrit will not let me. Please use latest sources.

It is apparent that whoever hits commit on this (me) is going to have to deal with many merge conflicts (does not kill this as long as the conflicts are trivial, but does make it clear that a complete sync has an up-front cost to get everyone inline). After rebase I will check a few trees to see how trivial it is (or not).

Patch-set: 4