Update patch set 1
Patch Set 1:

This fix shouldn't be necessary. The IPv6 gateway is configured by the Neighborhood Discovery Protocol (NDP), when the interface comes up it should send a Router Solicitation asking what gateway to use and the emulator should then respond with a Router Advertisement indicating that the guest should use fe80::2.

This is probably what happens on first boot, then something goes wrong when turning off flight mode that prevents this process from taking place.

I'm currently investigating a very similar issue where the emulator slirp code is dropping some packets, including Router Advertisements, because its packet queue ends up in an inconsistent state. I wonder if the issue you are seeing is the same problem and this is a workaround that forces the gateway to be set instead of relying on the Router Advertisements.

I would suggest holding off on merging this until I've discovered the issue with the dropped packets which I believe is the core issue here.

Patch-set: 1
Reviewer: Bjoern Johansson <1087048@85c56323-6fa9-3386-8a01-6480fb634889>