Update patch set 1
Patch Set 1:

Oh, yeah that's actually something we changed as part of introducing WiFi but that change is not in every image yet. In that case we can either make this change in images that don't have WiFi yet or we can cherry-pick the RIL change that disables ethernet when flight mode is turned on. I would prefer the cherry-picking since then we won't have to remember to remove this once WiFi is ported but I'm OK with making this change too if you prefer that.

Which one would you rather do?

Patch-set: 1
Reviewer: Bjoern Johansson <1087048@85c56323-6fa9-3386-8a01-6480fb634889>