Update patch set 1
Patch Set 1:

I'm not actually sure how real devices behave here. The main reason I'd rather not have both solutions is because NDP is more flexible in case the emulator decides to use another gateway some day. I don't think it's a huge deal though as the NDP messages will probably overwrite this gateway anyway, it might just be confusing to have both, for example if someone changes this and expects the gateway to be different but then the NDP message will overwrite it anyway. Or if the NDP message doesn't actually overwrite it and then the image is stuck with a non-working gateway.

I'll see if I can figure out what real devices do.

Patch-set: 1
Reviewer: Bjoern Johansson <1087048@85c56323-6fa9-3386-8a01-6480fb634889>