Update patch set 1
Patch Set 1:

So the modem firmware will intercept the network traffic and take the NDP router advertisements data and use that to respond to this RIL message? The emulator obviously doesn't do that, the RAs just reach the kernel and possibly netd and they will set it up. I guess they might do both.

Yeah I don't think hardcoding it here is a big problem, as you say we're unlikely to change it. If you prefer making this change I'm fine with that too and then we can re-evaluate once we start merging WiFi to more images. Let me know if you prefer this change and I'll +2 it.

Patch-set: 1
Reviewer: Bjoern Johansson <1087048@85c56323-6fa9-3386-8a01-6480fb634889>