Update patch set 1
Patch Set 1:

It seems to work for cellular too with the wifi solution. The way that's set up is using NDP for both radio and wifi. And since the framework is able to set up the correct IPv6 routes in that situation it seems that that's a solution that works.

Patch-set: 1
Reviewer: Bjoern Johansson <1087048@85c56323-6fa9-3386-8a01-6480fb634889>