Update patch set 1
Patch Set 1: Code-Review+2

Yes, that's mostly my understanding as well. The only thing that's not correct (as I understand it) is that IPv6 works because the kernel sets up the route automatically because of NDP. The IPv6 route is still set up the same way as for IPv4 with specific routing tables instead of using the default routing table. My understanding is that IPv6 very deliberately works together with NDP.

I belive that for a real device when you enable flight mode the RIL will instruct the modem to shut down based on a RIL_REQUEST_DEACTIVATE_DATA_CALL request. This in turn will bring down the network interface that RIL has set up as a way to send network traffic over the modem.  Then when it enables cellular data again the interface will be brought up and thus automatically send a router solicitation. The change to the RIL simply emulates this behavior by disabling or enabling the interface.

That change is here: https://googleplex-android-review.git.corp.google.com/#/c/2102269/

Now that I look at it though I see that's interconnected with the WiFi changes in a way that will require manual fixing before the cherry-pick goes through.

Perhaps your change is better as a workaround. Since the WiFi changes also touches on the gateway settings there will be a conflict once we start porting WiFi and that will be a good reminder to deal with it then.

Sorry about all this back and forth, I now think we should go with your change for now.

Patch-set: 1
Reviewer: Bjoern Johansson <1087048@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2